move silent syntax and annotate to own packages
