+++ projects/batfish/src/main/java/org/batfish/grammar/cisco_asa/AsaControlPlaneExtractor.java
@@ -413,7 +413,6 @@ import org.batfish.datamodel.vendor_family.cisco.User;
 import org.batfish.grammar.BatfishCombinedParser;
 import org.batfish.grammar.BatfishParseTreeWalker;
 import org.batfish.grammar.ControlPlaneExtractor;
-import org.batfish.grammar.SilentSyntaxCollection;
 import org.batfish.grammar.SilentSyntaxListener;
 import org.batfish.grammar.UnrecognizedLineToken;
 import org.batfish.grammar.cisco_asa.AsaParser.Aaa_accountingContext;
@@ -996,6 +995,7 @@ import org.batfish.grammar.cisco_asa.AsaParser.Vrfd_route_targetContext;
 import org.batfish.grammar.cisco_asa.AsaParser.Vrrp_interfaceContext;
 import org.batfish.grammar.cisco_asa.AsaParser.Wccp_idContext;
 import org.batfish.grammar.cisco_asa.AsaParser.Zp_service_policy_inspectContext;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
 import org.batfish.representation.cisco_asa.AccessListAddressSpecifier;
 import org.batfish.representation.cisco_asa.AccessListServiceSpecifier;
 import org.batfish.representation.cisco_asa.AsaConfiguration;
