+++ projects/batfish/src/main/java/org/batfish/grammar/cisco_xr/CiscoXrControlPlaneExtractor.java
@@ -354,7 +354,6 @@ import org.batfish.datamodel.vendor_family.cisco_xr.User;
 import org.batfish.grammar.BatfishCombinedParser;
 import org.batfish.grammar.BatfishParseTreeWalker;
 import org.batfish.grammar.ControlPlaneExtractor;
-import org.batfish.grammar.SilentSyntaxCollection;
 import org.batfish.grammar.SilentSyntaxListener;
 import org.batfish.grammar.UnrecognizedLineToken;
 import org.batfish.grammar.cisco_xr.CiscoXrParser.Aaa_accountingContext;
@@ -878,6 +877,7 @@ import org.batfish.grammar.cisco_xr.CiscoXrParser.Vrf_block_rb_stanzaContext;
 import org.batfish.grammar.cisco_xr.CiscoXrParser.Vrf_descriptionContext;
 import org.batfish.grammar.cisco_xr.CiscoXrParser.Vrf_nameContext;
 import org.batfish.grammar.cisco_xr.CiscoXrParser.Vrrp_interfaceContext;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
 import org.batfish.representation.cisco_xr.AccessListAddressSpecifier;
 import org.batfish.representation.cisco_xr.AccessListServiceSpecifier;
 import org.batfish.representation.cisco_xr.AsPathSet;
