+++ projects/batfish/src/main/java/org/batfish/grammar/cumulus_nclu/CumulusNcluControlPlaneExtractor.java
@@ -8,7 +8,7 @@ import org.batfish.common.Warnings;
 import org.batfish.grammar.BatfishParseTreeWalker;
 import org.batfish.grammar.ControlPlaneExtractor;
 import org.batfish.grammar.ImplementedRules;
-import org.batfish.grammar.SilentSyntaxCollection;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
 import org.batfish.representation.cumulus_nclu.CumulusNcluConfiguration;
 import org.batfish.vendor.VendorConfiguration;
