+++ projects/batfish/src/main/java/org/batfish/grammar/flatvyos/FlatVyosControlPlaneExtractor.java
@@ -18,7 +18,6 @@ import org.batfish.datamodel.Prefix;
 import org.batfish.grammar.BatfishCombinedParser;
 import org.batfish.grammar.BatfishParseTreeWalker;
 import org.batfish.grammar.ControlPlaneExtractor;
-import org.batfish.grammar.SilentSyntaxCollection;
 import org.batfish.grammar.SilentSyntaxListener;
 import org.batfish.grammar.flatvyos.FlatVyosParser.Bnt_nexthop_selfContext;
 import org.batfish.grammar.flatvyos.FlatVyosParser.Bnt_remote_asContext;
@@ -75,6 +74,7 @@ import org.batfish.grammar.flatvyos.FlatVyosParser.S_protocols_bgpContext;
 import org.batfish.grammar.flatvyos.FlatVyosParser.Srt_next_hopContext;
 import org.batfish.grammar.flatvyos.FlatVyosParser.St_host_nameContext;
 import org.batfish.grammar.flatvyos.FlatVyosParser.Statict_routeContext;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
 import org.batfish.representation.vyos.BgpNeighbor;
 import org.batfish.representation.vyos.BgpProcess;
 import org.batfish.representation.vyos.EspGroup;
