+++ projects/batfish/src/main/java/org/batfish/grammar/palo_alto/PaloAltoConfigurationBuilder.java
@@ -108,7 +108,6 @@ import org.batfish.datamodel.Prefix;
 import org.batfish.datamodel.RoutingProtocol;
 import org.batfish.datamodel.SubRange;
 import org.batfish.grammar.BatfishCombinedParser;
-import org.batfish.grammar.SilentSyntaxCollection;
 import org.batfish.grammar.SilentSyntaxListener;
 import org.batfish.grammar.UnrecognizedLineToken;
 import org.batfish.grammar.palo_alto.PaloAltoParser.Bgp_asnContext;
@@ -379,6 +378,7 @@ import org.batfish.grammar.palo_alto.PaloAltoParser.Vrrt_metricContext;
 import org.batfish.grammar.palo_alto.PaloAltoParser.Vrrtn_ipContext;
 import org.batfish.grammar.palo_alto.PaloAltoParser.Vrrtn_next_vrContext;
 import org.batfish.grammar.palo_alto.PaloAltoParser.Yes_or_noContext;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
 import org.batfish.representation.palo_alto.AddressGroup;
 import org.batfish.representation.palo_alto.AddressObject;
 import org.batfish.representation.palo_alto.AddressPrefix;
