+++ projects/batfish/src/main/java/org/batfish/grammar/palo_alto/PaloAltoControlPlaneExtractor.java
@@ -9,7 +9,7 @@ import org.batfish.common.Warnings;
 import org.batfish.grammar.BatfishParseTreeWalker;
 import org.batfish.grammar.ControlPlaneExtractor;
 import org.batfish.grammar.ImplementedRules;
-import org.batfish.grammar.SilentSyntaxCollection;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
 import org.batfish.representation.palo_alto.PaloAltoConfiguration;
 import org.batfish.vendor.VendorConfiguration;
