+++ projects/batfish/src/test/java/org/batfish/grammar/cisco_nxos/NxosOspfTest.java
@@ -39,7 +39,7 @@ import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.Vrf;
 import org.batfish.datamodel.matchers.OspfAreaMatchers;
 import org.batfish.datamodel.ospf.OspfArea;
-import org.batfish.grammar.SilentSyntaxCollection;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
 import org.batfish.main.Batfish;
 import org.batfish.main.BatfishTestUtils;
 import org.batfish.representation.cisco_nxos.CiscoNxosConfiguration;
