+++ projects/batfish/src/test/java/org/batfish/grammar/f5_bigip_imish/F5BigipImishGrammarTest.java
@@ -103,9 +103,9 @@ import org.batfish.datamodel.routing_policy.Environment;
 import org.batfish.datamodel.routing_policy.Environment.Direction;
 import org.batfish.datamodel.routing_policy.Result;
 import org.batfish.datamodel.routing_policy.RoutingPolicy;
-import org.batfish.grammar.SilentSyntaxCollection;
 import org.batfish.grammar.f5_bigip_structured.F5BigipStructuredCombinedParser;
 import org.batfish.grammar.f5_bigip_structured.F5BigipStructuredControlPlaneExtractor;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
 import org.batfish.main.Batfish;
 import org.batfish.main.BatfishTestUtils;
 import org.batfish.main.TestrigText;
