+++ projects/batfish/src/test/java/org/batfish/grammar/iptables/IpTablesGrammarTest.java
@@ -15,7 +15,7 @@ import org.apache.commons.lang3.SerializationUtils;
 import org.batfish.common.BatfishLogger;
 import org.batfish.common.Warnings;
 import org.batfish.config.Settings;
-import org.batfish.grammar.SilentSyntaxCollection;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
 import org.batfish.main.Batfish;
 import org.batfish.representation.iptables.IptablesVendorConfiguration;
 import org.junit.Test;
