+++ projects/batfish/src/test/java/org/batfish/job/ParseVendorConfigurationResultTest.java
@@ -15,8 +15,8 @@ import org.batfish.common.Warnings.ParseWarning;
 import org.batfish.datamodel.ConfigurationFormat;
 import org.batfish.datamodel.answers.ParseStatus;
 import org.batfish.datamodel.answers.ParseVendorConfigurationAnswerElement;
-import org.batfish.grammar.SilentSyntaxCollection;
-import org.batfish.grammar.SilentSyntaxCollection.SilentSyntaxElem;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection;
+import org.batfish.grammar.silent_syntax.SilentSyntaxCollection.SilentSyntaxElem;
 import org.batfish.representation.cisco.CiscoConfiguration;
 import org.batfish.vendor.VendorConfiguration;
 import org.junit.Test;
